Normalize multimedia licenses where possible.
Verbatim license term values in multimedia extensions are normalized to an
appropriate URI, including for non-GBIF-supported/old licenses.

No effort is made where the rights holder is mixed in with the license,
or to presume licenses for media by straightforward associatedMedia terms.

See https://github.com/gbif/occurrence/issues/78, this achieves ① and ②.